general api package cleanup
prior to centralizing api pacakges, the naming conventions
and location/ownership of helper functions were scattered
across multiple places.  This change focuses on moving
code around so that sets of code are clearly owned, easily
identified, and deduplicated throughout the package.

Some minor logic changes within, primarily around the funcs,
which were moved to common packages.  The rest of the
changes are code movement.
