+++ src/cmd/getM365/exchange/get_item.go
@@ -15,7 +15,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/alcionai/corso/src/cli/utils"
-	"github.com/alcionai/corso/src/internal/common"
+	"github.com/alcionai/corso/src/internal/common/str"
 	"github.com/alcionai/corso/src/pkg/account"
 	"github.com/alcionai/corso/src/pkg/backup/details"
 	"github.com/alcionai/corso/src/pkg/credentials"
@@ -54,7 +54,7 @@ func handleExchangeCmd(cmd *cobra.Command, args []string) error {
 		return nil
 	}
 
-	tid := common.First(tenant, os.Getenv(account.AzureTenantID))
+	tid := str.First(tenant, os.Getenv(account.AzureTenantID))
 
 	ctx := clues.Add(
 		cmd.Context(),
