--- src/internal/common/slices.go
@@ -1,23 +1,0 @@
-package common
-
-// TODO: can be replaced with slices.Contains()
-func ContainsString(super []string, sub string) bool {
-	for _, s := range super {
-		if s == sub {
-			return true
-		}
-	}
-
-	return false
-}
-
-// First returns the first non-zero valued string
-func First(vs ...string) string {
-	for _, v := range vs {
-		if len(v) > 0 {
-			return v
-		}
-	}
-
-	return ""
-}
