+++ src/pkg/services/m365/api/client.go
@@ -75,17 +75,3 @@ func newLargeItemService(creds account.M365Config) (*graph.Service, error) {
 
 	return a, nil
 }
-
-// ---------------------------------------------------------------------------
-// common types and consts
-// ---------------------------------------------------------------------------
-
-// DeltaUpdate holds the results of a current delta token.  It normally
-// gets produced when aggregating the addition and removal of items in
-// a delta-queryable folder.
-type DeltaUpdate struct {
-	// the deltaLink itself
-	URL string
-	// true if the old delta was marked as invalid
-	Reset bool
-}
