+++ src/pkg/services/m365/api/consts.go
@@ -1,0 +1,3 @@
+package api
+
+const maxPageSize = int32(999)
